Refactor component imports and restructure Search,Search-offchain component for improved organization
