+++ components/search/NFTGallery.tsx
