+++ components/search/WalletInfo.tsx
